requirements.txt
fastapi
uvicorn
strawberry-graphql
databases
sqlalchemy